+++ app/src/main/java/com/app/reelsapp/authentication/presentation/screen/register/RegisterViewModel.kt
@@ -1,9 +1,8 @@
 package com.app.reelsapp.authentication.presentation.screen.register
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.app.reelsapp.authentication.domain.AuthenticationRepository
+import com.app.reelsapp.authentication.data.repository.AuthenticationRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jakarta.inject.Inject
 import kotlinx.coroutines.Dispatchers
